Add RSpec testing framework to the engine
Configured a basic dummy app to test the engine, and set up the tests to
be ran from the root of the spec.

Remove tracking of log and db on dummy App

Cleaned the repo by removing test databases and logs from the repo,
added them to git ignore.

updated engine version in preparation to merge

Removed sqlite dependency